Update styles.css
Step 8 Passed

Now that your external CSS file is set up, you can start styling the page.

As a reminder, here's how to target a paragraph element and align it to the right:

p {
  text-align: right;
}

Create a new CSS rule that targets the h1 element, and set its text-align property to center.