Update styles.css
Step 46 Passed

A very common way to apply color to an element with CSS is with hexadecimal or hex values. While hex values sound complicated, they're really just another form of RGB values.

Hex color values start with a # character and take six characters from 0-9 and A-F. The first pair of characters represent red, the second pair represent green, and the third pair represent blue. For example, #4B5320.

In the .green class selector, set the background-color property to a hex color code with the values 00 for red, FF for green, and 00 blue.